add stable version with modpack flow

+++ services/cache/modpack_cache.go
@@ -3,6 +3,7 @@ package modpack_cache
 import (
 	"fmt"
 	"os"
+	"path/filepath"
 	"time"
 
 	config "github.com/brutalzinn/boberto-modpack-api/configs"
@@ -10,26 +11,44 @@ import (
 	"github.com/patrickmn/go-cache"
 )
 
-var modpack_cache = cache.New(1*time.Second, 5*time.Second)
+var modpack_cache = cache.New(5*time.Minute, 10*time.Minute)
 
 func New() {
 	config := config.GetConfig()
 	modpack_cache.OnEvicted(func(s string, modpackExpire any) {
 		oldModPack := modpackExpire.(*models.Modpack)
-		finalPath := fmt.Sprintf("%s/%s", config.PublicPath, oldModPack.NormalizedName)
-		fmt.Printf("trying to remove %s", finalPath)
-		os.RemoveAll(finalPath)
+		if oldModPack.Status == models.Aborted ||
+			oldModPack.Status == models.Canceled ||
+			oldModPack.Status == models.Error {
+			finalPath := filepath.Join(config.PublicPath, oldModPack.NormalizedName)
+			fmt.Printf("trying to remove %s", finalPath)
+			os.RemoveAll(finalPath)
+		}
+		fmt.Printf("Delete cache for %s", s)
 	})
 }
 
-func GetModpackCacheById(id string) *models.Modpack {
-	if modpack, found := modpack_cache.Get(id); found {
-		return modpack.(*models.Modpack)
-	}
-	return nil
+func SetStatus(id string, status models.ModPackStatus) models.Modpack {
+	modpack, _ := GetModpackCacheById(id)
+	modpack.Status = status
+	modpack_cache.SetDefault(id, modpack)
+	return modpack
 }
 
-func CreateModpackCache(modpack *models.Modpack) string {
+func GetStatus(id string) models.ModPackStatus {
+	modpack, _ := GetModpackCacheById(id)
+	status := modpack.Status
+	return status
+}
+
+func GetModpackCacheById(id string) (modpack models.Modpack, found bool) {
+	if modpack, found := modpack_cache.Get(id); found {
+		return modpack.(models.Modpack), true
+	}
+	return models.Modpack{}, false
+}
+
+func CreateModpackCache(modpack models.Modpack) string {
 	modpack_cache.Set(modpack.Id, modpack, cache.DefaultExpiration)
 	return modpack.Id
 }
